+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Profile rest service dockerized
 - Image update and fileDescriptor moved to pii from non-pii
 - Test profiles building block rest service deployed in AWS server
+- Add searching by mobile app version feature to App Config API
+- Add secretKeys to App Config
 
 ### Changed
 - Update events schema to remove endDate as required field and add createdBy and createDate.